# backend/docs/swagger.yaml
openapi: 3.0.1
info:
  title: suito
  contact: {}
  version: "1.0"
servers:
- url: //localhost:8009/api/v1
paths:
...
  /transactions:
    get:
      tags:
      - suito.transactions
      summary: List transactions
      description: List transactions.
      operationId: listTransactions
      parameters:
      - name: yearMonth
        in: query
        description: yearMonth
        required: true
        schema:
          type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransactionsRes'
        500:
          description: Unknown Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuitoError'
  /transactions/months:
    get:
      tags:
      - suito.transactions
      summary: List transaction months
      description: トランザクションが発生した月(YYYY-MM)の一覧を取得します.
      operationId: transactionMonths
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionMonthsRes'
        500:
          description: Unknown Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuitoError'
components:
  schemas:
...
    ListTransactionsRes:
      required:
      - transactions
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...
    SuitoError:
      type: object
      properties:
        ErrCode:
          type: string
        message:
          type: string
    Transaction:
      required:
      - amount
      - id
      - localDate
      - title
      - type
      type: object
      properties:
        amount:
          type: integer
        id:
          type: string
        localDate:
          type: string
        title:
          type: string
        type:
          type: integer
    TransactionMonthsRes:
      required:
      - yearMonths
      type: object
      properties:
        yearMonths:
          type: array
          items:
            type: string
...